BREAKING CHANGE: Fix bug in import order preventing FancyPage overwrite
The FancyPage model from django-fancypages was imported *before*
the overwriting FancyPage model in this app. As a result the page would
not reference the Oscar category properly. This is fixed now by making
sure that the models used from django-fancypages are imported **after**
defining the FancyPage model in this app.

**Warnging:** This bug fix will break existing project using
django-oscar-fancypages as I had to regenerate the initial migration.
Since this is an import issue and not an actual model change, I didn't
see a way around this via migrations. Sorry for any inconvenience.